feat: add "sql" generic database connector.
Adds a new connector, "sql", that allows Dex to connect to an external
database (presently, sqlite3, PostgreSQL and MySQL are supported) and
obtain user and group information from that database.

In order to securely support passwords, this connector supports bcrypt
hashed passwords.  It doesn't support any other type of hashed
password (yet), and it doesn't support plaintext password.

Also added a set of tests for it.

Signed-off-by: Alastair Houghton <alastair@alastairs-place.net>