ci: archive the bin/dex binary.
While we build a Docker image, we also want to be able to construct
an RPM, and for that we need a non-Alpine build (which is what we've
got by the end of this job), but we need it in the cloud-hub-dex
job rather than here.  So, archive it from this job so that we can
fetch it from the cloud-hub-dex job.